Add --run command

+++ JpegToWord/DocCreator.cs
@@ -1,4 +1,3 @@
-using System;
 using Newtonsoft.Json;
 using Spire.Doc;
 using Spire.Doc.Documents;
@@ -25,7 +24,7 @@ namespace JpegToWord
                 image.HorizontalAlignment = ShapeHorizontalAlignment.Center;
                 Image img = Image.FromFile(arg);
                 image.Width = 500;
-                image.Height = 500 * ((float)img.Height / (float)img.Width);
+                image.Height = 500 * (img.Height / (float)img.Width);
             }
 
             Paragraph footer = section.AddParagraph();
@@ -77,7 +76,7 @@ namespace JpegToWord
                 image.HorizontalAlignment = ShapeHorizontalAlignment.Center;
                 Image img = Image.FromFile(arg);
                 image.Width = 500;
-                image.Height = 500 * ((float)img.Height / (float)img.Width);
+                image.Height = 500 * (img.Height / (float)img.Width);
             }
 
             Paragraph footer = section.AddParagraph();
